requirements.txt
streamlit
nibabel
numpy
matplotlib
pandas
python-dotenv
groq